UI Re-done.
Re-done the UI from the basic set-up that had the HealthBar up the top left of the screen to now it having a Health, Oxygen, Ammo in clip and ammo in reserves set up. (Note none of the Icon's are currently set up to the players stats eg The players health could be 2 but the Icon would still show 100. Will be added in next push.)